+++ concert/devices/__init__.py
@@ -1,1 +1,3 @@
 """__init__.py"""
+
+from concert.devices import base
