+++ concert/experiments/base.py
@@ -7,23 +7,14 @@ import asyncio
 import functools
 import inspect
 import logging
-import os
 import time
-import json
 
-import concert.devices.base
 from concert.coroutines.base import background, broadcast, start
 from concert.coroutines.sinks import count
 from concert.progressbar import wrap_iterable
 from concert.base import (
-    check,
-    Parameter,
-    Selection,
-    StateError,
-    RunnableParameterizable, TargetAccessError
+    RunnableParameterizable,
 )
-from concert.helpers import get_basename
-from concert.loghandler import AsyncLoggingHandlerCloser
 from functools import partial
 
 
@@ -219,12 +210,6 @@ class Acquisition(RunnableParameterizable):
                     await self.producer.unregister_endpoint(consumer.addon.endpoint)
                     await consumer.addon.disconnect_endpoint()
 
-    @background
-    @check(source=['standby', 'error', 'cancelled'], target=['standby', 'cancelled'])
-    async def __call__(self):
-        self._run_awaitable = self._run()
-        await self._run_awaitable
-
     def __repr__(self):
         return "Acquisition({})".format(self.name)
 
@@ -264,154 +249,15 @@ class Experiment(RunnableParameterizable):
 
     """
 
-    iteration = Parameter()
-    separate_scans = Parameter()
-    name_fmt = Parameter()
-    current_name = Parameter(help="Name of the current iteration")
-    log_level = Selection(['critical', 'error', 'warning', 'info', 'debug'])
-    log_devices_at_start = Parameter()
-    log_devices_at_finish = Parameter()
+
 
     async def __ainit__(self, *, acquisitions, walker=None, separate_scans=True, name_fmt='scan_{:>04}', **kwargs):
+
         self._acquisitions = []
         for acquisition in acquisitions:
             self.add(acquisition)
-        self.walker = walker
-        self._separate_scans = separate_scans
-        self._name_fmt = name_fmt
-        self._current_name = ""
-        self._iteration = 0
-        self.log = LOG
-        self._devices_to_log = {}
-        self._devices_to_log_optional = {}
-        self._log_devices_at_start = None
-        self._log_devices_at_finish = None
-        self.ready_to_prepare_next_sample = asyncio.Event()
-        await super().__ainit__(**kwargs)
-        await self.set_log_devices_at_start(True)
-        await self.set_log_devices_at_finish(True)
-
-        if separate_scans and walker:
-            # The data is not supposed to be overwritten, so find an iteration which
-            # hasn't been used yet
-            while await self.walker.exists(self._name_fmt.format(self._iteration)):
-                self._iteration += 1
-
-    async def _set_log_devices_at_start(self, log):
-        self._log_devices_at_start = bool(log)
-
-    async def _get_log_devices_at_start(self):
-        return self._log_devices_at_start
-
-    async def _set_log_devices_at_finish(self, log):
-        self._log_devices_at_finish = bool(log)
-
-    async def _get_log_devices_at_finish(self):
-        return self._log_devices_at_finish
-
-    def add_device_to_log(self, name: str, device: concert.devices.base.Device, optional=False):
-        """
-        Add a device to log.
-
-        :param name: Name of the device
-        :param device: Device to log
-        :param optional: If True, an exception when trying to log the device will not cause an
-            error.
-        """
-        if optional:
-            self._devices_to_log_optional[name] = device
-        else:
-            self._devices_to_log[name] = device
-
-    async def _prepare_metadata_str(self) -> str:
-        """Prepares the experiment metadata to be written to file. It is
-        a dictionary which potentially encapsulates one or more dictionary
-        objects.
-        """
-        metadata = {}
-        exp_params = {}
-        for param in self:
-            exp_params[param.name] = str(await param.get())
-        metadata["experiment"] = exp_params
-        for name, device in self._devices_to_log.items():
-            device_data = {}
-            for param in device:
-                device_data[param.name] = {}
-                device_data[param.name]["value"] = str(await param.get())
-                try:
-                    device_data[param.name]["target"] = str(await param.get_target())
-                except TargetAccessError:
-                    pass
-                try:
-                    device_data[param.name]["lower_limit"] = str(await param.get_lower())
-                    device_data[param.name]["upper_limit"] = str(await param.get_upper())
-                    device_data[param.name]["lower_user_limit"] = str(await param.get_lower_user())
-                    device_data[param.name]["upper_user_limit"] = str(await param.get_upper_user())
-                    device_data[param.name]["lower_external_limit"] = str(await param.get_lower_external())
-                    device_data[param.name]["upper_external_limit"] = str(await param.get_upper_external())
-                except AttributeError:
-                    pass
-            metadata[name] = device_data
-
-        for name, device in self._devices_to_log_optional.items():
-            device_data = {}
-            for param in device:
-                try:
-                    device_data[param.name] = {}
-                    device_data[param.name]["value"] = str(await param.get())
-                    try:
-                        device_data[param.name]["target"] = str(await param.get_target())
-                    except TargetAccessError:
-                        pass
-                    try:
-                        device_data[param.name]["lower_limit"] = str(await param.get_lower())
-                        device_data[param.name]["upper_limit"] = str(await param.get_upper())
-                        device_data[param.name]["lower_user_limit"] = str(await param.get_lower_user())
-                        device_data[param.name]["upper_user_limit"] = str(await param.get_upper_user())
-                        device_data[param.name]["lower_external_limit"] = str(await param.get_lower_external())
-                        device_data[param.name]["upper_external_limit"] = str(await param.get_upper_external())
-                    except AttributeError:
-                        pass
-                except Exception as e:
-                    self.log.info(f"Error while logging optional device {name}")
-                    self.log.info(e)
-            metadata[name] = device_data
-        return json.dumps(metadata, indent=4)
-
-    async def _get_iteration(self):
-        return self._iteration
-
-    async def _set_iteration(self, iteration):
-        self._iteration = iteration
-
-    async def _get_current_name(self):
-        return self._current_name
-
-    async def _get_separate_scans(self):
-        return self._separate_scans
-
-    async def _set_separate_scans(self, separate_scans):
-        self._separate_scans = separate_scans
-
-    async def _get_name_fmt(self):
-        return self._name_fmt
-
-    async def _set_name_fmt(self, fmt):
-        self._name_fmt = fmt
-
-    async def _get_log_level(self):
-        return logging.getLevelName(self.log.getEffectiveLevel()).lower()
-
-    async def _set_log_level(self, level):
-        self.log.setLevel(level.upper())
-
-    async def prepare(self):
-        """Gets executed before every experiment run."""
-        pass
-
-    async def finish(self):
-        """Gets executed after every experiment run."""
-        pass
+        await super().__ainit__(walker=walker, separate_scans=separate_scans,name_fmt=name_fmt, **kwargs)
+        await self.set_log_file_prefix("experiment")
 
     async def attach(self, addon):
         """Attach *addon* to all acquisitions."""
@@ -488,62 +334,11 @@ class Experiment(RunnableParameterizable):
         for acq in wrap_iterable(self._acquisitions):
             if await self.get_state() != 'running':
                 break
-            await acq()
+            await acq.run()
 
     @background
     async def _run(self):
-        self.ready_to_prepare_next_sample.clear()
-        start_time = time.time()
-        handler = None
-        iteration = await self.get_iteration()
-        separate_scans = await self.get_separate_scans()
-
-        if self.walker:
-            if separate_scans:
-                await self.walker.descend((await self.get_name_fmt()).format(iteration))
-            if os.path.exists(await self.walker.get_current()):
-                handler: AsyncLoggingHandlerCloser = await self.walker.register_logger(
-                    logger_name=self.__class__.__name__,
-                    log_level=logging.NOTSET,
-                    file_name="experiment.log"
-                )
-                self.log.addHandler(handler)
-                if await self.get_log_devices_at_start():
-                    exp_metadata: str = await self._prepare_metadata_str()
-                    await self.walker.log_to_json(payload=exp_metadata,
-                                                  filename="experiment_start.json")
-            self._current_name = get_basename(await self.walker.get_current())
-        self.log.info(await self.info_table)
-        for name, device in self._devices_to_log.items():
-            self.log.info(f"Device {name}:")
-            self.log.info(await device.info_table)
-        LOG.debug('Experiment iteration %d start', iteration)
-
-        try:
-            await self.prepare()
-            await self.acquire()
-        finally:
-            try:
-                await self.finish()
-                if self.walker:
-                    if await self.get_log_devices_at_finish():
-                        exp_metadata: str = await self._prepare_metadata_str()
-                        await self.walker.log_to_json(payload=exp_metadata,
-                                                      filename="experiment_finish.json")
-            except Exception as e:
-                LOG.warning(f"Error `{e}' while finalizing experiment")
-                raise StateError('error', msg=str(e))
-            finally:
-                self.ready_to_prepare_next_sample.set()
-                if separate_scans and self.walker:
-                    await self.walker.ascend()
-                LOG.debug('Experiment iteration %d duration: %.2f s',
-                          iteration, time.time() - start_time)
-                if handler:
-                    await handler.aclose()
-                    self.log.removeHandler(handler)
-                await self.set_iteration(iteration + 1)
-
+        await self.acquire()
 
 class AcquisitionError(Exception):
     """Acquisition-related exceptions."""
